src/prime: improve 'get_all_primes_below_n'.
Check the return of the method 'fill_tab_prime_bool'.

+++ src/math/prime.h
@@ -13,7 +13,7 @@ typedef struct prime_factor_t {
 } prime_factor_t;
 
 bool is_prime(unsigned long n);
-unsigned int
+long
 get_all_primes_below_n(unsigned long lim, unsigned int size_tab_out,
                        unsigned long *out);
 int get_all_n_first_primes(unsigned long count_asked,
